* 4.4:
  Fix composer on appveyor
  [PropertyAccess] Fix typo in PropertyAccessor::readProperty() DocBlock
  [VarExporter] Fix exporting objects with readonly properties
  [ExpressionLanguage] Fix matches when the regexp is not valid
  [Messenger] Add mysql indexes back and work around deadlocks using soft-delete
  [Validator] Fix File constraint invalid max size exception message
  [Console] Fix exit status on uncaught exception with negative code

+++ src/Symfony/Component/Messenger/Bridge/Doctrine/Transport/Connection.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Messenger\Bridge\Doctrine\Transport;
 
 use Doctrine\DBAL\Connection as DBALConnection;
+use Doctrine\DBAL\Driver\Exception as DriverException;
 use Doctrine\DBAL\Driver\Result as DriverResult;
 use Doctrine\DBAL\Exception as DBALException;
 use Doctrine\DBAL\Exception\TableNotFoundException;
@@ -153,6 +154,14 @@ class Connection implements ResetInterface
 
     public function get(): ?array
     {
+        if ($this->driverConnection->getDatabasePlatform() instanceof MySQLPlatform) {
+            try {
+                $this->driverConnection->delete($this->configuration['table_name'], ['delivered_at' => '9999-12-31']);
+            } catch (DriverException $e) {
+                // Ignore the exception
+            }
+        }
+
         get:
         $this->driverConnection->beginTransaction();
         try {
@@ -224,6 +233,10 @@ class Connection implements ResetInterface
     public function ack(string $id): bool
     {
         try {
+            if ($this->driverConnection->getDatabasePlatform() instanceof MySQLPlatform) {
+                return $this->driverConnection->update($this->configuration['table_name'], ['delivered_at' => '9999-12-31'], ['id' => $id]) > 0;
+            }
+
             return $this->driverConnection->delete($this->configuration['table_name'], ['id' => $id]) > 0;
         } catch (DBALException $exception) {
             throw new TransportException($exception->getMessage(), 0, $exception);
@@ -233,6 +246,10 @@ class Connection implements ResetInterface
     public function reject(string $id): bool
     {
         try {
+            if ($this->driverConnection->getDatabasePlatform() instanceof MySQLPlatform) {
+                return $this->driverConnection->update($this->configuration['table_name'], ['delivered_at' => '9999-12-31'], ['id' => $id]) > 0;
+            }
+
             return $this->driverConnection->delete($this->configuration['table_name'], ['id' => $id]) > 0;
         } catch (DBALException $exception) {
             throw new TransportException($exception->getMessage(), 0, $exception);
@@ -404,6 +421,7 @@ class Connection implements ResetInterface
         $table->addColumn('headers', Types::TEXT)
             ->setNotnull(true);
         $table->addColumn('queue_name', Types::STRING)
+            ->setLength(190) // MySQL 5.6 only supports 191 characters on an indexed column in utf8mb4 mode
             ->setNotnull(true);
         $table->addColumn('created_at', Types::DATETIME_MUTABLE)
             ->setNotnull(true);
@@ -412,11 +430,8 @@ class Connection implements ResetInterface
         $table->addColumn('delivered_at', Types::DATETIME_MUTABLE)
             ->setNotnull(false);
         $table->setPrimaryKey(['id']);
-        // No indices on queue_name and available_at on MySQL to prevent deadlock issues when running multiple consumers.
-        if (!$this->driverConnection->getDatabasePlatform() instanceof MySQLPlatform) {
-            $table->addIndex(['queue_name']);
-            $table->addIndex(['available_at']);
-        }
+        $table->addIndex(['queue_name']);
+        $table->addIndex(['available_at']);
         $table->addIndex(['delivered_at']);
     }
